Rstpm2-package(Rstpm2)	R Documentation

Flexible parametric survival models.

Description

The package implements the stpm2 models from Stata. Such models use a flexible parametric formulation for survival models, using natural splines to model the log-cumulative hazard. Model predictions are rich, allowing for direct estimation of the hazard, survival, hazard ratios, hazard differences and survival differences. The models allow for time-varying effects, left truncation and relative survival.

The R implementation departs from the Stata implementation, using the ns() function, which is based on a projection of B-splines, rather than using truncated power splines as per Stata.

Details

Package: 	 Rstpm2

Type: 	 Package

Version: 	 1.0

Date: 	 2011-07-06

License: 	 GPL-2

LazyLoad: 	 yes

Depends: 	 methods, bbmle

Imports: 	 splines, survival, stats, graphics

The package exports the stpm2 object, which inherits from the mle2 object from the bbmle package. Methods are specified for the stpm2 object, including predict and plot methods.

Author(s)

Mark Clements and Paul Lambert.

Maintainer: <mark.clements@ki.se>

See Also

stpm2

Examples


data(brcancer)
summary(fit <- stpm2(Surv(rectime,censrec==1)~hormon,data=brcancer,df=3))
summary(fit.tvc <- stpm2(Surv(rectime,censrec==1)~hormon,data=brcancer,df=3,
                     tvc=list(hormon=3)))
anova(fit,fit.tvc)
plot(fit.tvc,newdata=data.frame(hormon=0),type="hr",var="hormon")

[Package Rstpm2 version 1.0 Index]