coef<- {rstpm2}	R Documentation

Generic method to update the coef in an object.

Description

Generic method to update the coef in an object.

Usage

coef(x) <- value

Arguments

x
object to be updated

value
value of the coefficient to be updated.

Details

This simple generic method is used for the numerical delta method.

Value

The updated object is returned.

Examples

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x, value) 
UseMethod("coef<-")

[Package rstpm2 version 1.7.1 Index]