eform.stpm2 {rstpm2}	R Documentation

S3 method for to provide exponentiated coefficents with confidence intervals.

Description

S3 method for to provide exponentiated coefficents with confidence intervals.

Usage

eform(object, ...)
## S3 method for class 'stpm2'
eform(object, parm, level = 0.95, method = c("Profile","Delta"),
    name = "exp(beta)", ...)
## Default S3 method:
eform(object, parm, level = 0.95, method =
c("Delta","Profile"), name = "exp(beta)", ...)

Arguments

object
regression object

parm
not currently used

level
significance level for the confidence interval

method
method for confidence interval estimation

name
name for the fitted value

...
other arguments

[Package rstpm2 version 1.7.1 Index]