grad {rstpm2}	R Documentation

gradient function (internal function)

Description

Numerical gradient for a function at a given value (internal).

Usage

grad(func, x, ..., method=c("fast","richardson"))

Arguments

func
Function taking a vector argument x (returns a vector of length>=1)

x
vector of arguments for where the gradient is wanted.

...
other arguments to the function

method
string argument to determine whether to use the fast two sided calculation or use a Richardson extrapolation.

Details

(func(x+delta,...)-func(x-delta,...))/(2 delta) where delta is the third root of the machine precision times pmax(1,abs(x)).

Value

A vector if func(x) has length 1, otherwise a matrix with rows for x and columns for func(x).

Author(s)

Mark Clements.

See Also

numDelta()

[Package rstpm2 version 1.7.1 Index]