gsm.control {rstpm2}	R Documentation

Defaults for the gsm call

Description

Set useful default and allow changes for the gsm call. This is meant to make the gsm call simpler.

Usage

gsm.control(parscale = 1, maxit = 300, optimiser = c("BFGS", "NelderMead"), trace = 0,
            nodes = 9, adaptive = TRUE, kappa.init = 1, maxkappa = 1000,
            suppressWarnings.coxph.frailty = TRUE, robust_initial = FALSE, bhazinit = 0.1,
            eps.init = 1e-5, use.gr = TRUE, penalty = c("logH", "h"), outer_optim = 1,
            reltol.search = 1e-10, reltol.final = 1e-10, reltol.outer = 1e-05,
            criterion = c("GCV", "BIC"), old.init=FALSE)

Arguments

parscale
numeric vector or scalar for the scaling of the parameter values; default 1

maxit
integer for the maximum number of iterations for the optimisation process

optimiser
which optimiser to use for the outer optimisation

trace
integer indicating the trace level for each optimiser

nodes
number of quadrature nodes

adaptive
logical for whether to use adaptive or non-adaptive quadrature, Default=TRUE.

kappa.init
initial value for the quadratic penalty for inequality constraints

eps.init
initial value for epsilon

maxkappa
double float value for the maximum value of the weight used in the constraint.

suppressWarnings.coxph.frailty
logical

robust_initial
Not currently documented.

bhazinit
Not currently documented.

use.gr
Logical for whether to use gradients.

penalty
Not currently documented.

outer_optim
Not currently documented.

reltol.search
Relative tolerance. Not currently documented.

reltol.final
Relative tolerance. Not currently documented.

reltol.outer
Relative tolerance. Not currently documented.

criterion
Not currently documented.

old.init
Use the initial values from version 1.6.8 and earlier, which uses the Shat_old function.

[Package rstpm2 version 1.7.1 Index]