stpm2(Rstpm2)	R Documentation

Usage


stpm2(formula, data, df = 3, logH.args = NULL, logH.formula = NULL, tvc = NULL, tvc.formula = NULL, bhazard = NULL, contrasts = NULL, ...)

Arguments

formula 	 a formula object, with the response on the left of a ~ operator, and the terms on the right. The response must be a survival object as returned by the Surv function. The terms should include linear terms for any time-varying coefficients. [required]

data 	 a data.frame in which to interpret the variables named in the formula argument. [at present: required]

df 	 an integer that describes the degrees of freedom for the ns function for modelling the baseline log-cumulative hazard (default=3).

logH.args 	 a list describing the arguments for the ns function for modelling the baseline log-cumulative hazard (default=NULL). Use this for changing the knot placement.

logH.formula 	 a formula for describing the baseline log-cumulative hazard function (default=NULL). Only one of df, logH.args or logH.formula is required.

tvc 	 a list with the names of the time-varying coefficients and the degrees of freedom (e.g. tvc=list(x=3) specifies x as a time-varying coefficient with 3 degrees of freedom).

tvc.formula 	 a formula for describing the time-varying coefficients. If a time-varying coefficient is being model, then only one of tvc and tvc.formula is required.

bhazard 	 a vector for the background hazard for relative survival estimation. At present, this does not use data. At present, this is required for all individuals - although it is only used at the event times.

contrasts 	 an optional list. See the contrasts.arg of model.matrix.default.

... 	 additional arguments to be passed to the mle2 .

Details

The implementation extends the mle2 object from the bbmle package.

Value

An stpm2-class object.

Examples


data(brcancer)
summary(fit <- stpm2(Surv(rectime,censrec==1)~hormon,data=brcancer,df=3))

## some predictions
head(predict(fit,se.fit=T,type="surv"))
head(predict(fit,se.fit=T,type="hazard"))

## some plots
plot(fit,newdata=data.frame(hormon=0),type="hazard")
plot(fit,newdata=data.frame(hormon=0),type="surv")

## time-varying coefficient
summary(fit.tvc <- stpm2(Surv(rectime,censrec==1)~hormon,data=brcancer,df=3,
                     tvc=list(hormon=3)))
anova(fit,fit.tvc) # compare with and without tvc

## some more plots
plot(fit.tvc,newdata=data.frame(hormon=0),type="hr",var="hormon")
                                        # no lines method: use add=TRUE
plot(fit.tvc,newdata=data.frame(hormon=1),type="hr",var="hormon",
     add=TRUE,ci=FALSE,line.col=2)

plot(fit.tvc,newdata=data.frame(hormon=0),type="sdiff",var="hormon")

plot(fit.tvc,newdata=data.frame(hormon=0),type="hdiff",var="hormon")

plot(fit.tvc,newdata=data.frame(hormon=0),type="hazard")
plot(fit.tvc,newdata=data.frame(hormon=1),type="hazard",line.col=2,ci=FALSE,add=TRUE)

## compare number of knots
hormon0 <- data.frame(hormon=0)
plot(fit,type="hazard",newdata=hormon0)
AIC(fit)
for (df in 4:6) {
    fit.new <- stpm2(Surv(rectime,censrec==1)~hormon,data=brcancer,df=df)
    plot(fit.new,type="hazard",newdata=hormon0,add=T,ci=F,line.col=df)
    print(AIC(fit.new))
}

## The function is currently defined as
function (formula, data = NULL, df = 3, logH.args = NULL, logH.formula = NULL, 
    tvc = NULL, tvc.formula = NULL, bhazard = NULL, contrasts = NULL, 
    ...) 
{
    cl <- match.call()
    mf <- match.call(expand.dots = FALSE)
    m <- match(c("formula", "data"), names(mf), 0L)
    mf <- mf[c(1L, m)]
    mf$drop.unused.levels <- TRUE
    mf[[1L]] <- as.name("model.frame")
    eventExpression <- lhs(formula)[[length(lhs(formula))]]
    delayed <- length(lhs(formula)) == 4
    timevar <- lhs(formula)[[if (delayed) 
        3
    else 2]]
    if (is.null(logH.formula) && is.null(logH.args)) 
        logH.args$df <- df
    if (is.null(logH.formula)) 
        logH.formula <- as.formula(call("~", as.call(c(quote(ns), 
            call("log", timevar), vector2call(logH.args)))))
    full.formula <- formula
    rhs(full.formula) <- rhs(formula) %call+% rhs(logH.formula)
    if (is.null(tvc.formula) && !is.null(tvc)) {
        tvc.formulas <- lapply(names(tvc), function(name) call(":", 
            as.name(name), as.call(c(quote(ns), call("log", timevar), 
                vector2call(list(df = tvc[[name]]))))))
        if (length(tvc.formulas) > 1) 
            tvc.formulas <- list(Reduce(`%call+%`, tvc.formulas))
        tvc.formula <- as.formula(call("~", tvc.formulas[[1]]))
    }
    if (!is.null(tvc.formula)) 
        rhs(full.formula) <- rhs(full.formula) %call+% rhs(tvc.formula)
    logHD.formula = replaceFormula(logH.formula, quote(ns), quote(nsDeriv))
    if (!is.null(tvc.formula)) {
        tvcD.formula <- replaceFormula(tvc.formula, quote(ns), 
            quote(nsDeriv))
        rhs(logHD.formula) <- logHD.formula[[2]] %call+% tvcD.formula[[2]]
    }
    mf$formula = full.formula
    mf$subset <- eventExpression
    mfd <- mf
    mf <- eval(mf, parent.frame())
    mt <- attr(mf, "terms")
    xlev <- .getXlevels(mt, mf)
    mfd[[2]] <- logHD.formula
    mfd <- eval(mfd, parent.frame())
    mtd <- attr(mfd, "terms")
    mfX <- model.frame(mt, data, xlev = xlev)
    if (!is.null(cl <- attr(mt, "dataClasses"))) 
        .checkMFClasses(cl, mfX)
    X <- model.matrix(mt, mfX, contrasts)
    mfXD <- model.frame(mtd, data, xlev = xlev)
    if (!is.null(cl <- attr(mtd, "dataClasses"))) 
        .checkMFClasses(cl, mfXD)
    XD <- model.matrix(mtd, mfXD, contrasts)[, -1, drop = FALSE]
    y <- model.extract(mfX, "response")
    if (!inherits(y, "Surv")) 
        stop("Response must be a survival object")
    type <- attr(y, "type")
    if (type != "right" && type != "counting") 
        stop(paste("stpm2 model doesn't support \"", type, "\" survival data", 
            sep = ""))
    event <- y[, ncol(y)] == 1
    time <- y[, ncol(y) - 1]
    coxph.obj <- coxph(formula, data = data)
    data$logHhat <- log(-log(Shat(coxph.obj)))
    lm.formula <- full.formula
    lhs(lm.formula) <- quote(logHhat)
    init <- coef(lm(lm.formula, data[event, ], contrasts = contrasts))
    indexXD <- (length(coef(coxph.obj)) + 2):ncol(X)
    bhazard <- if (is.null(bhazard)) 
        0
    else bhazard[event]
    if (delayed && any(y[, 1] > 0)) {
        data2 <- data[y[, 1] > 0, , drop = FALSE]
        y.call <- lhs(full.formula)
        delayed.formula <- full.formula
        lhs(delayed.formula) <- NULL
        data2[[deparse(y.call[[3]])]] <- data2[[deparse(y.call[[2]])]]
        mfX2 <- model.frame(mt, data2, xlev = xlev)
        if (!is.null(cl <- attr(mt, "dataClasses"))) 
            .checkMFClasses(cl, mfX2)
        X2 <- model.matrix(mt, mfX2, contrasts)[, -1, drop = FALSE]
        negll <- function(beta) {
            eta <- X %*% beta
            eta2 <- X2 %*% beta
            h <- (XD[event, ] %*% beta[indexXD]) * exp(eta[event])/time[event] + 
                bhazard
            h[h < 0] <- 1e-100
            ll <- sum(log(h)) + sum(exp(eta2)) - sum(exp(eta))
            return(-ll)
        }
    }
    else {
        negll <- function(beta) {
            eta <- X %*% beta
            h <- (XD[event, ] %*% beta[indexXD]) * exp(eta[event])/time[event] + 
                bhazard
            h[h < 0] <- 1e-100
            ll <- sum(log(h)) - sum(exp(eta))
            return(-ll)
        }
    }
    parnames(negll) <- names(init)
    mle2 <- mle2(negll, init, vecpar = T, ...)
    new("stpm2", call = mle2@call, call.orig = mle2@call, coef = mle2@coef, 
        fullcoef = mle2@fullcoef, vcov = mle2@vcov, min = mle2@min, 
        details = mle2@details, minuslogl = mle2@minuslogl, method = mle2@method, 
        data = data, formula = mle2@formula, optimizer = "optim", 
        xlevels = .getXlevels(mt, mf), contrasts = attr(X, "contrasts"), 
        Call = cl, terms = mt, model.frame = mf, x = X, xd = XD, 
        termsd = mtd, y = y)
  }

[Package Rstpm2 version 1.0 Index]