predict-methods(Rstpm2)	R Documentation

Predicted values for an stpm2 fit

Description

Given an stpm2 fit and an optional list of new data, return predictions

Usage



    ## S4 method for signature 'stpm2':
    predict(object, newdata=NULL,
               type=c("surv","cumhaz","hazard","hr","sdiff","hdiff"),
               grid=FALSE,seqLength=300,
               se.fit=FALSE,link=NULL,exposed=incrVar(var),var,...)

Arguments

object 	 an stpm2 object

newdata 	 optional list of new data (required if type in ("hr","sdiff","hdiff")). For type in ("hr","sdiff","hdiff"), this defines the unexposed newdata. This can be combined with grid to get a regular set of event times (i.e. newdata would not include the event times).

type 	 specify the type of prediction:

"surv"

survival probabilities

"cumhaz"

cumulative hazard

"hazard"

hazard

"hr"

hazard ratio

"sdiff"

survival difference

"hdiff"

hazard difference

grid 	 whether to merge newdata with a regular sequence of event times (default=FALSE)

seqLength 	 length of the sequence used when grid=TRUE

se.fit 	 whether to calculate confidence intervals (default=FALSE)

link 	 allows a different link for the confidence interval calculation (default=NULL)

exposed 	 a function that takes newdata and returns a transformed data-frame for those exposed

var 	 specify the variable name or names for the exposed/unexposed (names are given as characters)

... 	 additional arguments (for generic compatibility)

Details

The confidence interval estimation is based on the delta method using numerical differentiation.

Value

A data-frame with components Estimate, lower and upper, with an attribute "newdata" for the newdata data-frame.

Methods

object= "stpm2"

an stpm2 fit

See Also

stpm2

Examples



## The function is currently defined as
setMethod("predict", "stpm2",
          function(object,newdata=NULL,type=c("surv","cumhaz","hazard","hr","sdiff","hdiff"),
                   grid=FALSE,seqLength=300,
                   se.fit=FALSE,link=NULL,exposed=incrVar(var),var,...)
  {
    ## exposed is a function that takes newdata and returns the revised newdata
    ## var is a string for a variable that defines a unit change in exposure
    local <-  function (object, newdata=NULL, type="surv", exposed)
      {
        tt <- object@terms 
        if (is.null(newdata)) {
          ##mm <- X <- model.matrix(object) # fails (missing timevar)
          X <- object@x
          XD <- object@xd
          y <- model.response(object@model.frame) 
          time <- as.vector(y[,ncol(y)-1])
        }
        else {
          Terms <- delete.response(tt)
          m <- model.frame(Terms, newdata, xlev = object@xlevels)
          if (!is.null(cl <- attr(Terms, "dataClasses"))) 
            .checkMFClasses(cl, m)
          X <- model.matrix(Terms, m, contrasts.arg = object@contrasts)
          resp <- attr(Terms, "variables")[attr(Terms, "response")]
          ## similarly for the derivatives
          if (type 
            ttd <- object@termsd
            TermsD <- delete.response(ttd)
            md <- model.frame(TermsD, newdata, xlev = object@xlevels)
            if (!is.null(cld <- attr(TermsD, "dataClasses"))) 
              .checkMFClasses(cld, md)
            XD <- model.matrix(TermsD, md, contrasts.arg = object@contrasts)[,-1,drop=FALSE]
            ## how to elegantly extract the time variable?
            timevar <- if (length(tt[[2]])==3) tt[[2]][[2]] else tt[[2]][[3]]
            time <- model.matrix(as.formula(call("~",timevar)),newdata)[,-1,drop=TRUE]
            ##
          }
          if (type 
            if (missing(exposed))
              stop("exposed needs to be specified for type in ('hr','sdiff','hdiff')")
            newdata2 <- exposed(newdata)
            m2 <- model.frame(Terms, newdata2, xlev = object@xlevels)
            if (!is.null(cl <- attr(Terms, "dataClasses"))) 
              .checkMFClasses(cl, m2)
            X2 <- model.matrix(Terms, m2, contrasts.arg = object@contrasts)
            md2 <- model.frame(TermsD, newdata2, xlev = object@xlevels)
            if (!is.null(cld <- attr(TermsD, "dataClasses"))) 
              .checkMFClasses(cld, md2)
            XD2 <- model.matrix(TermsD, md2, contrasts.arg = object@contrasts)[,-1,drop=FALSE]
          }
        }
        beta <- coef(object)
        cumHaz = exp(X 
        Sigma = vcov(object)
        if (type=="cumhaz") { # delayed entry?
          return(cumHaz)
        }
        if (type=="surv") { # delayed entry?
          return(exp(-cumHaz))
        }
        if (type=="sdiff")
          return(exp(-exp(X2 
        if (type=="hazard") {
          betaXD <- beta[(ncol(X)-ncol(XD)+1):ncol(X)]
          return((XD 
        }
        if (type=="hdiff") {
          betaXD <- beta[(ncol(X)-ncol(XD)+1):ncol(X)]
          return((XD2 
        }
        if (type=="hr") {
          cumHazRatio = exp((X2 - X) 
          betaXD <- beta[(ncol(X)-ncol(XD)+1):ncol(X)]
          return((XD2 
        }
      

[Package Rstpm2 version 1.1.1 Index]